Add files via upload
2020-04-09
Dear Dr. Ng:
The Mask parts are donated by: Richard Smeenk, Agile Manufacturing, Inc. (Agile) at the request of Brian Read. The masks are known as Stopgap Face Mask (SFM) by the National Institute of Health (NIH):
This device is intended to be use only for the duration of the public health emergency related to COVID-19. The decision to implement this device should be made with careful consideration and under the consultation of the corresponding institution’s occupational health and infection control departments.
 
The Stopgap Face Mask was designed to be a surgical mask intended to provide liquid barrier protection for medical purposes… This mask should be used within the CDC guidelines for Strategies for Optimizing the Supply of N95 Respirators: Crisis/Alternate Strategies ,
 
This device is not suitable protection against airborne exposures and should not be used as a replacement for a N95 mask, PAPR device, or any other respirator device. This supplementary face mask DOES NOT MEET REQUIREMENTS FOR AIRBORNE PRECAUTIONS and SHOULD NOT BE USED DURING AEROSOL GENERATING PROCEDURES. The supplementary mask should not be used in a clinical setting where the infection risk level through inhalation exposure is high
 
The Face Masks have been made in DuraForm PA nylon. A regulatory information sheet for DuraForm is enclosed. These nylon parts are steam sterilizable many times. Please note Agile is not an ISO 10993 facility, sterilize at your facility.
 
As you can see from the certification the DuraForm nylon is suitable for skin contact. These parts have been tumbled smooth to improve the surface finish for comfort. This Mask is designed for use without an edge seal.
 
Let us know how we may be of further service. Stay well.
Regards,
Richard Smeenk
Agile Manufacturing
P: 905.852.0794, F: 905.852.1546, C: 905.852.0923